# docs/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "FO76 DB",
    "description": "Локальный API утилиты FO76 DB: каталог предметов, инвентарь, роллы, события, настройки.\n\n**Доступ.** С этого компьютера (127.0.0.1, без прокси) токен не нужен. Из сети, если в настройках задан токен, передавайте\n`Authorization: Bearer <токен>` (или cookie `fo76db_token`, или `?token=` один раз). Без токена — 401.\nИзменяющие запросы с чужого `Origin` без Bearer отклоняются (защита от CSRF). Для браузерного фронтенда на другом адресе\nдобавьте его в `cors_origins` в `config.toml`. Запись путей, настроек сети, конфига IOM и подобное — только с этого компьютера.",
    "version": "2026.10.5"
  },
  "paths": {
...
    "/api/tags": {
      "get": {
        "summary": "Api Tags",
        "operationId": "api_tags_api_tags_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          }
        },
        "tags": [
          "tags"
        ]
      },
      "post": {
        "summary": "Api Tag Save",
        "operationId": "api_tag_save_api_tags_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Body_api_tag_save_api_tags_post"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "tags": [
          "tags"
        ]
      }
    },
    "/api/tags/{tag_id}": {
      "delete": {
        "summary": "Api Tag Delete",
        "operationId": "api_tag_delete_api_tags__tag_id__delete",
        "parameters": [
          {
            "name": "tag_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Tag Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "tags": [
          "tags"
        ]
      }
    },
    "/api/tags/apply": {
      "post": {
        "summary": "Api Tag Apply",
        "operationId": "api_tag_apply_api_tags_apply_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Body_api_tag_apply_api_tags_apply_post"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "tags": [
          "tags"
        ]
      }
    },
    "/api/display_name": {
      "post": {
        "summary": "Api Display Name",
        "operationId": "api_display_name_api_display_name_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Body_api_display_name_api_display_name_post"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "tags": [
          "display_name"
        ]
      }
    },
    "/api/wishlist": {
      "post": {
        "summary": "Api Wishlist",
        "operationId": "api_wishlist_api_wishlist_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Body_api_wishlist_api_wishlist_post"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "tags": [
          "wishlist"
        ]
      }
    },
    "/api/saved": {
      "get": {
        "summary": "Api Saved",
        "operationId": "api_saved_api_saved_get",
        "parameters": [
          {
            "name": "page",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Page"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "tags": [
          "saved"
        ]
      },
      "post": {
        "summary": "Api Saved Add",
        "operationId": "api_saved_add_api_saved_post",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Body_api_saved_add_api_saved_post"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "tags": [
          "saved"
        ]
      }
    },
    "/api/saved/{sid}": {
      "delete": {
        "summary": "Api Saved Delete",
        "operationId": "api_saved_delete_api_saved__sid__delete",
        "parameters": [
          {
            "name": "sid",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Sid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "tags": [
          "saved"
        ]
      }
    },
    "/api/item/{formid}/wiki": {
      "get": {
        "summary": "Api Item Wiki",
        "description": "Описание и «где взять» со страниц предмета: русская fandom, английская fandom или fallout.wiki. Кэш в `wiki_obtain`.",
        "operationId": "api_item_wiki_api_item__formid__wiki_get",
        "parameters": [
          {
            "name": "formid",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Formid"
            }
          },
          {
            "name": "refresh",
            "in": "query",
            "required": false,
            "schema": {
              "type": "boolean",
              "default": false,
              "title": "Refresh"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "tags": [
          "item"
        ]
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "Body_api_display_name_api_display_name_post": {
        "properties": {
          "key": {
            "type": "string",
            "title": "Key"
          },
          "name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Name"
          }
        },
        "type": "object",
        "required": [
          "key"
        ],
        "title": "Body_api_display_name_api_display_name_post"
      },
...
      "Body_api_saved_add_api_saved_post": {
        "properties": {
          "page": {
            "type": "string",
            "title": "Page"
          },
          "name": {
            "type": "string",
            "title": "Name"
          },
          "params": {
            "type": "string",
            "title": "Params"
          }
        },
        "type": "object",
        "required": [
          "page",
          "name",
          "params"
        ],
        "title": "Body_api_saved_add_api_saved_post"
      },
...
      "Body_api_tag_apply_api_tags_apply_post": {
        "properties": {
          "tag_id": {
            "type": "integer",
            "title": "Tag Id"
          },
          "keys": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Keys"
          },
          "on": {
            "type": "boolean",
            "title": "On"
          }
        },
        "type": "object",
        "required": [
          "tag_id",
          "keys",
          "on"
        ],
        "title": "Body_api_tag_apply_api_tags_apply_post"
      },
      "Body_api_tag_save_api_tags_post": {
        "properties": {
          "name": {
            "type": "string",
            "title": "Name"
          },
          "color": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Color"
          },
          "icon": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Icon"
          },
          "id": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Id"
          }
        },
        "type": "object",
        "required": [
          "name"
        ],
        "title": "Body_api_tag_save_api_tags_post"
      },
      "Body_api_wishlist_api_wishlist_post": {
        "properties": {
          "formid": {
            "type": "string",
            "title": "Formid"
          },
          "wish": {
            "type": "boolean",
            "title": "Wish"
          },
          "note": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Note"
          }
        },
        "type": "object",
        "required": [
          "formid",
          "wish"
        ],
        "title": "Body_api_wishlist_api_wishlist_post"
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    },
    "securitySchemes": {
      "bearer": {
        "type": "http",
        "scheme": "bearer",
        "description": "token из config.toml / «Настройки → Доступ по сети»"
      }
    }
  },
...
}